thegraph-hyper: Split code into separate modules
