+++ thegraph-hyper/src/lib.rs
@@ -7,6 +7,12 @@ extern crate slog;
 extern crate thegraph;
 extern crate tokio;
 
+mod request;
+mod response;
 mod server;
+mod service;
 
+pub use self::request::GraphQLRequest;
+pub use self::response::GraphQLResponse;
 pub use self::server::HyperGraphQLServer;
+pub use self::service::{GraphQLService, GraphQLServiceResponse};
